third_party/blink/web_tests/external/wpt/trusted-types/should-trusted-type-policy-creation-be-blocked-by-csp-002-expected.txt

This is a testharness.js-based test.
[FAIL] invalid tt-policy-name name "política"
  assert_true: createPolicy() should throw a TypeError. expected true got false
[FAIL] directive "trusted-type _TTP1_%09_TTP2_%0A_TTP3_%0C_TTP4_%0D_TTP5_%20_TTP6_" (required-ascii-whitespace)
  assert_equals: expected null but got object "TypeError: Failed to execute 'createPolicy' on 'TrustedTypePolicyFactory': Policy "_TTP3_" disallowed."
Harness: the test ran to completion.
